Split showcase data into opt-in seed
